フォーマット修正

+++ RebuildUs/Modules/CustomOptions/CustomTaskOption.cs
@@ -11,12 +11,23 @@ internal sealed class CustomTasksOption
                              (int commonTasks, int shortTasks, int longTasks) defaultValue,
                              CustomOption parent = null)
     {
-        _commonTaskOption = CustomOption.Normal(ids.commonId, type, TrKey.NumCommonTask, (float)defaultValue.commonTasks, 0f, 4f, 1f, parent);
-        _longTaskOption = CustomOption.Normal(ids.longId, type, TrKey.NumLongTask, (float)defaultValue.longTasks, 0f, 23f, 1f, parent);
-        _shortTaskOption = CustomOption.Normal(ids.shortId, type, TrKey.NumShortTask, (float)defaultValue.shortTasks, 0f, 15f, 1f, parent);
+        _commonTaskOption = CustomOption.Normal(ids.commonId, type, TrKey.NumCommonTask, defaultValue.commonTasks, 0f, 4f, 1f, parent);
+        _longTaskOption = CustomOption.Normal(ids.longId, type, TrKey.NumLongTask, defaultValue.longTasks, 0f, 23f, 1f, parent);
+        _shortTaskOption = CustomOption.Normal(ids.shortId, type, TrKey.NumShortTask, defaultValue.shortTasks, 0f, 15f, 1f, parent);
     }
 
-    internal int CommonTasksNum { get => Mathf.RoundToInt(_commonTaskOption.GetFloat()); }
-    internal int ShortTasksNum { get => Mathf.RoundToInt(_shortTaskOption.GetFloat()); }
-    internal int LongTasksNum { get => Mathf.RoundToInt(_longTaskOption.GetFloat()); }
+    internal int CommonTasksNum
+    {
+        get => Mathf.RoundToInt(_commonTaskOption.GetFloat());
+    }
+
+    internal int ShortTasksNum
+    {
+        get => Mathf.RoundToInt(_shortTaskOption.GetFloat());
+    }
+
+    internal int LongTasksNum
+    {
+        get => Mathf.RoundToInt(_longTaskOption.GetFloat());
+    }
 }
